Audit wave 2: the file-local remainder
Fourteen agents, one per file so each had a sole writer, working the verified
findings the account-boundary wave did not cover.

Capture: the untaken-loot batch survived the account boundary and was flushed
into the NEXT account's journal (and pushed under its token); the screenshot
PNG encode ran synchronously on the game render thread; serverLootKeys grew
without bound for server-only loot; a group-mate's despawned loot inside our
own kill window could be booked as ours; KILL_DURATION missed the lowercase
in-sentence phrasings ToA/ToB/Colosseum use, so those PBs never landed; and
the no-KC screenshot value floor was dead code.

Journal + panel: a corrupt journal is set aside rather than silently
overwritten; rename-follow no longer accumulates onto a record already filed
under the new name (it is preserved as a sidecar and this account's own
journal wins its name); the History tab parses the calendar spine on the
executor instead of re-reading the whole append-only file on the EDT per
click; a floor-only stats section showed ONE row labelled by the first floor
key but valued at the SUM of all of them (Prayer read "Bones buried 15,190"
when 3,112 were buried); the History empty state stopped promising a cloud
adoption the teardown removed.

Correctness elsewhere: patchesRaked minted a rake for any weeds ENTERING the
inventory (bank, GE, trade, ground pickup); totalXpGained dropped the next
delta after every region cross; the cloud-sync toggle rebased the journal off
the client thread, where a concurrent fold could double it; prayerVerbMigrated
was client-global rather than per-account; the one downward read is now
bounded and its adopted figures clamped before they reach the system of
record; and the Loot Tracker is declared a dependency (its LootReceived event
is the only path chest and casket loot takes) with a panel nudge if disabled.

AccountBoundaryTest pins the store contract wave 1's guards rest on.

+++ src/main/java/chronicle/HistoryLog.java
@@ -40,8 +40,10 @@ class HistoryLog
 	// enough to make the day-rollover append fire exactly once.
 	private volatile String lastAppendedDate;
 
-	/** Append today's closing baseline. Call at login-load, day rollover, logout. */
-	synchronized void append(File dir, String rsn, Map<String, Integer> skills,
+	/** Append today's closing baseline. Call at login-load, day rollover, logout.
+	 *  Both maps are long-valued: total xp outgrows an int, and the reader takes
+	 *  every number back as a long. */
+	synchronized void append(File dir, String rsn, Map<String, Long> skills,
 		Map<String, Long> counters)
 	{
 		if (rsn == null || rsn.isEmpty())
